refactor(shared): extract text-rendering pure logic (line-height, warp, effects)
Move the framework-agnostic text CSS-builder logic into pptx-viewer-shared.
React behavior is unchanged (it now consumes the shared helpers); the new
modules also unblock future Vue/Angular enrichment.

Done across all bindings that had a copy:
- render/mathml-sanitize.ts: sanitizeMathMl (React + Vue; Angular keeps DomSanitizer).
- render/text-warp.ts: groupIntoParagraphs warp-paragraph splitter (React + Vue
  + Angular).
- render/text-style-helpers.ts: resolveLineHeight (React + Vue, and now wired
  into Angular getTextBlockStyle, closing its missing-line-height gap).

Extracted to shared + repointed React (neutral Record<string,string|number> CSS
records; enables a later Vue/Angular pass):
- text-style-helpers: vertical-text writing-mode/orientation/direction +
  computeAutoFitTextStyle.
- text-decoration.ts: resolveUnderlineDecorationStyle (16 OOXML underline styles).
- text-paragraph-style.ts: resolveParagraphRtl/Align + resolveCssTextAlign.
- text-field-substitution.ts: substituteFieldText/resolveFieldDateText.
- text-fill.ts / text-effects.ts / text-effects-3d.ts: text fill, shadow, glow,
  blur, HSL, reflection, alpha, 3D extrusion CSS builders.

shared now depends on dompurify (already used by React/Vue); added to shared and
to Angular (which vendors shared at build).

Follow-up (documented): wiring the B-group helpers into the Vue/Angular text
renderers (vertical text, underline variants, field substitution, text effects)
needs template changes with visual-regression risk, deferred to a targeted pass.

Verified: shared build + 2714 tests; React 6289 (unchanged), Vue 1421, Angular
1925 tests; all four typecheck clean.
