+++ packages/react/src/viewer/utils/text-utils.tsx
@@ -1,4 +1,11 @@
 import { PptxElement, TextSegment, TextStyle, hasTextProperties } from 'pptx-viewer-core';
+import {
+	computeAutoFitTextStyle,
+	resolveLineHeight,
+	toCssTextOrientation,
+	toCssVerticalDirection,
+	toCssWritingMode,
+} from 'pptx-viewer-shared';
 import React from 'react';
 
 import {
@@ -10,6 +17,16 @@ import {
 } from '../constants';
 import { cloneTextStyle } from './clone';
 import { normalizeHexColor } from './color';
+
+// Vertical-text writing-mode helpers + line-height + auto-fit scaling now live
+// in pptx-viewer-shared (render/text-style-helpers). Re-exported here so
+// existing React import paths (`./text-utils`) keep working.
+export {
+	toCssWritingMode,
+	toCssTextOrientation,
+	toCssVerticalDirection,
+	isVerticalTextDirection,
+} from 'pptx-viewer-shared';
 
 export type ListMode = 'none' | 'bullet' | 'number';
 
@@ -171,170 +188,16 @@ export function getTextStyleForElement(
 			? { whiteSpace: 'nowrap' as const, overflow: 'visible' as const }
 			: {}),
 		// Auto-fit: use OOXML-provided fontScale/lnSpcReduction when available,
-		// otherwise fall back to heuristic estimation.
-		...(element.textStyle?.autoFit && hasTextProperties(element)
-			? (() => {
-					const baseFontSize = element.textStyle?.fontSize || DEFAULT_TEXT_FONT_SIZE;
-					const result: React.CSSProperties = {};
-
-					// normAutofit with explicit fontScale: use the exact percentage
-					if (
-						element.textStyle?.autoFitFontScale !== undefined &&
-						element.textStyle.autoFitFontScale > 0 &&
-						element.textStyle.autoFitFontScale < 1
-					) {
-						result.fontSize = Math.max(
-							6,
-							Math.round(baseFontSize * element.textStyle.autoFitFontScale),
-						);
-					} else if (element.textStyle?.autoFitMode !== 'normal') {
-						// spAutoFit (shrink): heuristic estimation
-						const textLength = (element.text ?? '').length;
-						const lineHeight = element.textStyle?.lineSpacingExactPt
-							? element.textStyle.lineSpacingExactPt / baseFontSize
-							: element.textStyle?.lineSpacing || (hasItalicRuns ? 1.35 : 1.25);
-						const approxCharsPerLine = Math.max(
-							1,
-							Math.floor(element.width / (baseFontSize * 0.6)),
-						);
-						const estimatedLines = Math.max(1, Math.ceil(textLength / approxCharsPerLine));
-						const requiredHeight = estimatedLines * baseFontSize * lineHeight;
-						const availableHeight = element.height - (bodyTop + bodyBottom);
-						if (requiredHeight > availableHeight && availableHeight > 0) {
-							const scale = Math.max(0.5, availableHeight / requiredHeight);
-							result.fontSize = Math.max(6, Math.round(baseFontSize * scale));
-						}
-					}
-
-					// normAutofit with lnSpcReduction: reduce line height
-					if (
-						element.textStyle?.autoFitLineSpacingReduction !== undefined &&
-						element.textStyle.autoFitLineSpacingReduction > 0
-					) {
-						const baseLineHeight =
-							typeof element.textStyle?.lineSpacing === 'number'
-								? element.textStyle.lineSpacing
-								: hasItalicRuns
-									? 1.35
-									: 1.25;
-						result.lineHeight =
-							baseLineHeight * (1 - element.textStyle.autoFitLineSpacingReduction);
-					}
-
-					return result;
-				})()
-			: {}),
+		// otherwise fall back to heuristic estimation. The pure font-scale maths
+		// now lives in pptx-viewer-shared (computeAutoFitTextStyle).
+		...computeAutoFitTextStyle({
+			textStyle: element.textStyle,
+			text: element.text ?? '',
+			width: element.width,
+			height: element.height,
+			bodyInsetVertical: bodyTop + bodyBottom,
+			hasItalicRuns,
+			defaultFontSize: DEFAULT_TEXT_FONT_SIZE,
+		}),
 	};
-}
-
-/**
- * Map a parsed `textDirection` value to the corresponding CSS `writing-mode`.
- *
- * | textDirection     | CSS writing-mode |
- * |-------------------|------------------|
- * | `"vertical"`      | `vertical-rl`    |
- * | `"eaVert"`        | `vertical-rl`    |
- * | `"wordArtVert"`   | `vertical-rl`    |
- * | `"wordArtVertRtl"`| `vertical-rl`    |
- * | `"vertical270"`   | `vertical-lr`    |
- * | `"mongolianVert"` | `vertical-lr`    |
- * | `"horizontal"`    | undefined        |
- */
-export function toCssWritingMode(
-	textDirection: TextStyle['textDirection'] | undefined,
-): React.CSSProperties['writingMode'] | undefined {
-	switch (textDirection) {
-		case 'vertical':
-		case 'eaVert':
-		case 'wordArtVert':
-		case 'wordArtVertRtl':
-			return 'vertical-rl';
-		case 'vertical270':
-		case 'mongolianVert':
-			return 'vertical-lr';
-		default:
-			return undefined;
-	}
-}
-
-/**
- * Resolve CSS `text-orientation` for vertical writing modes.
- *
- * | textDirection     | CSS text-orientation |
- * |-------------------|----------------------|
- * | `"vertical"`      | `mixed`              |
- * | `"eaVert"`        | `mixed`              |
- * | `"vertical270"`   | `mixed`              |
- * | `"wordArtVert"`   | `upright`            |
- * | `"wordArtVertRtl"`| `mixed`              |
- * | `"mongolianVert"` | `mixed`              |
- * | `"horizontal"`    | undefined            |
- *
- * - `"vertical"` / `"eaVert"`: CJK glyphs stay upright, Latin rotated (`mixed`).
- * - `"vertical270"`: text rotated 270deg, all glyphs rotated (`mixed`).
- * - `"wordArtVert"`: all glyphs rendered upright, stacked vertically (`upright`).
- * - `"wordArtVertRtl"`: same as vertical-rl with RTL direction (`mixed`).
- * - `"mongolianVert"`: Mongolian vertical, left-to-right columns (`mixed`).
- */
-export function toCssTextOrientation(
-	textDirection: TextStyle['textDirection'] | undefined,
-): React.CSSProperties['textOrientation'] | undefined {
-	switch (textDirection) {
-		case 'vertical':
-		case 'eaVert':
-		case 'vertical270':
-		case 'wordArtVertRtl':
-		case 'mongolianVert':
-			return 'mixed';
-		case 'wordArtVert':
-			return 'upright';
-		default:
-			return undefined;
-	}
-}
-
-/**
- * Resolve CSS `direction` override for vertical text modes that require RTL.
- *
- * Only `"wordArtVertRtl"` requires explicit `direction: rtl`.
- */
-export function toCssVerticalDirection(
-	textDirection: TextStyle['textDirection'] | undefined,
-): React.CSSProperties['direction'] | undefined {
-	if (textDirection === 'wordArtVertRtl') {
-		return 'rtl';
-	}
-	return undefined;
-}
-
-/**
- * Check whether a textDirection value represents any vertical writing mode.
- */
-export function isVerticalTextDirection(
-	textDirection: TextStyle['textDirection'] | undefined,
-): boolean {
-	return (
-		textDirection === 'vertical' ||
-		textDirection === 'vertical270' ||
-		textDirection === 'eaVert' ||
-		textDirection === 'wordArtVert' ||
-		textDirection === 'wordArtVertRtl' ||
-		textDirection === 'mongolianVert'
-	);
-}
-
-/**
- * Resolve CSS `line-height` from TextStyle.
- * If `lineSpacingExactPt` is set (exact point mode from spcPts), return a fixed `Xpt` string.
- * If `lineSpacing` is set (proportional mode from spcPct), return the multiplier.
- * Otherwise use a sensible default.
- */
-function resolveLineHeight(
-	textStyle: TextStyle | undefined,
-	hasItalicRuns: boolean,
-): string | number {
-	if (typeof textStyle?.lineSpacingExactPt === 'number' && textStyle.lineSpacingExactPt > 0) {
-		return `${textStyle.lineSpacingExactPt}pt`;
-	}
-	return textStyle?.lineSpacing || (hasItalicRuns ? 1.35 : 1.25);
 }
